requirements.txt
Flask
google-cloud-pubsub
firebase-admin